日時をPOSTする時や画面に表示する時のフォーマット用jsファイルを作成
- formatForDisplay関数の引数にはすでにISO形式の日時が入るためsliceメソッドのみで処理
- ISO形式にフォーマットするタイミングを修正
